Certs: refactor key generation script
This refactors the script to create a client keypair aswell (since the process is the same as for the server the code has been moved to a function). Also subject alternative names can now be given as an argument and the locations where resulting keystores are put are now relative to the new project structure. Of course formatting and codestructure has also been cleaned up.